chore: Reorganize input generators into separate class
