+++ packages/graphback/src/input/ContextTypes.ts
@@ -1,4 +1,4 @@
-import { GraphQLGeneratorConfig } from './GraphQLGeneratorConfig';
+import { GraphQLGeneratorConfig } from '../GraphQLGeneratorConfig';
 
 export const INTERFACE_TYPE_DEFINITION = 'InterfaceTypeDefinition';
 export const OBJECT_TYPE_DEFINITION = 'ObjectTypeDefinition';
