+++ packages/graphback/src/input/directives.ts
@@ -1,3 +1,5 @@
+
+// TODO split directives into individual plugins
 const directives = `directive @OneToMany(field: String) on FIELD_DEFINITION
 directive @OneToOne(field: String) on FIELD_DEFINITION
 directive @create(enable: Boolean) on OBJECT
